Fixed shutdown bug in HoneyBadgerService
Here I added another command line option to honey_badger;
the timeout for reading packets off the wire. Instead of blocking
we now timeout and this gives us a chance to catch a "shutdown signal".

I've added a Close() method to both the pcap packet logger and the tcp attack logger
so that they can do cleanup such as flush buffers and stop goroutines.

Added Close() to Connection... which simply calls Close() on the two above mentioned loggers.
Connection's TCP FSM calls close when it enters the TCP_CLOSED state.

I've also given the ConnTracker a Close() method which it uses to close ALL the
Connection objects in it's hashmap. This only get's called when honey_badger wants
to shutdown (user control-c).